requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.16
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
openpyxl>=3.1.0
pandas>=2.2.0
aiofiles>=23.2.0
python-dotenv>=1.0.0
itsdangerous>=2.2.0
